Use NODE_ENV instead of a custom env.js file
It's the standard in the node community to use NODE_ENV as the
environment variable. In addition, it's not recommended to have a file
that can change that we don't want to change checked into the repo. If
that number is inadvertently changed it could cause issues, although
there may be safety checks around it that I'm unaware of.

Therefore, remove the env.js file and replace it with NODE_ENV. This
changed the deployment script, and the api.js file. Default to "1" as
the CI environment.

Add error checking in the api.js file to verify that the environment set
is recognised.

Also update the README to explain the environments a little bit more, as
on first read not everyone may be familiar with "CI" and "UAT" as terms.
And include "local" in the environment possibilities.